+++ src/main/kotlin/step2/consoleView/inputValidator/FourBasicOperationsValidatorFacade.kt
@@ -1,10 +1,10 @@
-package step2.inputValidator
+package step2.consoleView.inputValidator
 
-import step2.inputValidator.Validators.validEmptyString
-import step2.inputValidator.Validators.validNull
-import step2.inputValidator.Validators.validNumberInExpression
-import step2.inputValidator.Validators.validOperatorInExpression
-import step2.inputValidator.Validators.validStartsAndEndsWithNumber
+import step2.consoleView.inputValidator.Validators.validEmptyString
+import step2.consoleView.inputValidator.Validators.validNull
+import step2.consoleView.inputValidator.Validators.validNumberInExpression
+import step2.consoleView.inputValidator.Validators.validOperatorInExpression
+import step2.consoleView.inputValidator.Validators.validStartsAndEndsWithNumber
 
 object FourBasicOperationsValidatorFacade {
     fun validate(input: List<String>?): Result<List<String>> =
